requirements.txt
pydbus
PyGObject
pyyaml
wand